fix unit test, share setup
